+++ flask_restplus/fields.py
@@ -42,14 +42,39 @@ def is_indexable_but_not_string(obj):
     return not hasattr(obj, "strip") and hasattr(obj, "__iter__")
 
 
-def get_value(key, obj, default=None):
-    '''Helper for pulling a keyed value off various types of objects'''
+def get_value(key, obj, default=None, dot_escape=False):
+    '''Helper for pulling a keyed value off various types of objects
+
+    :param bool dot_escape: Allow escaping of '.' character in field names to
+        indicate non-nested property access
+
+    >>> data = {'a': 'foo', b: {'c': 'bar', 'd.e': 'baz'}}}
+    >>> get_value('a', data)
+    'foo'
+
+    >>> get_value('b.c', data)
+    'bar'
+
+    >>> get_value('x', data, default='foobar')
+    'foobar'
+
+    >>> get_value('b.d\.e', data, dot_escape=True)
+    'baz'
+    '''
     if isinstance(key, int):
         return _get_value_for_key(key, obj, default)
     elif callable(key):
         return key(obj)
     else:
-        return _get_value_for_keys(key.split('.'), obj, default)
+        keys = (
+            [
+                k.replace("\.", ".")
+                for k in re.split(r"(?<!\\)\.", key)
+            ]
+            if dot_escape
+            else key.split(".")
+        )
+        return _get_value_for_keys(keys, obj, default)
 
 
 def _get_value_for_keys(keys, obj, default):
@@ -104,6 +129,8 @@ class Raw(object):
     :param bool readonly: Is the field read only ? (for documentation purpose)
     :param example: An optional data example (for documentation purpose)
     :param callable mask: An optional mask function to be applied to output
+    :param bool dot_escape: Allow escaping of '.' character in field names to
+        indicate non-nested property access
     '''
     #: The JSON/Swagger schema type
     __schema_type__ = 'object'
@@ -113,7 +140,8 @@ class Raw(object):
     __schema_example__ = None
 
     def __init__(self, default=None, attribute=None, title=None, description=None,
-                 required=None, readonly=None, example=None, mask=None, **kwargs):
+                 required=None, readonly=None, example=None, mask=None,
+                 dot_escape=False, **kwargs):
         self.attribute = attribute
         self.default = default
         self.title = title
@@ -122,6 +150,7 @@ class Raw(object):
         self.readonly = readonly
         self.example = example or self.__schema_example__
         self.mask = mask
+        self.dot_escape = dot_escape
 
     def format(self, value):
         '''
@@ -151,7 +180,11 @@ class Raw(object):
         :raises MarshallingError: In case of formatting problem
         '''
 
-        value = get_value(key if self.attribute is None else self.attribute, obj)
+        value = get_value(
+            key if self.attribute is None else self.attribute,
+            obj,
+            dot_escape=self.dot_escape
+        )
 
         if value is None:
             default = self._v('default')